Week_7/twitter_proj/etl/requirements_etl.txt
pymongo==3.11.0
SQLAlchemy==1.3.13
psycopg2-binary==2.8.5
